requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
Flask>=2.0.0

# 自动优化依赖
selenium>=4.0.0
webdriver-manager>=3.8.0
torch>=2.0.0
torchvision>=0.15.0